+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added type validation (without support for `__view_body__`)
 - Patched query strings on app testing
 - Added tests for query and body parameters
+- Patched body parameters
+- Documented type validation
+- Patched bodies with testing
 
 ## [1.0.0-alpha3] - 2023-09-9
 - Patched header responses